video_editing/montage_from_photos/requirements.txt
moviepy==1.0.3
Pillow==8.4.0
tqdm==4.62.3
